backend/requirements.txt
# backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
pymongo==3.12.0
authlib==0.15.4
PyJWT==2.1.0
requests==2.26.0
python-jose==3.3.0
werkzeug==2.0.1